Include bitmap memory estimate in AppWidgetServiceImpl dumpsys output
This change adds a new field "view_bitmap_memory" to the
AppWidgetServiceImpl dumpsys output, which holds the estimate of bitmap
memory used by a widget held in memory by the service.

To support this, a new method RemoteViews.estimateTotalBitmapMemoryUsage
is added, which includes memory from both bitmap and Icon actions.

Also updates RemoteViews.estimateMemoryUsage to return a long. The
int sum can be overflowed to beat the memory check in
AppWidgetServiceImpl.

Bug: 377356756
Test: RemoteViewsTest#estimateMemoryUsage_*
Flag: EXEMPT bugfix

+++ services/core/java/com/android/server/notification/NotificationManagerService.java
@@ -8683,7 +8683,7 @@ public class NotificationManagerService extends SystemService {
         if (contentView == null) {
             return false;
         }
-        final int contentViewSize = contentView.estimateMemoryUsage();
+        final long contentViewSize = contentView.estimateMemoryUsage();
         if (contentViewSize > mWarnRemoteViewsSizeBytes
                 && contentViewSize < mStripRemoteViewsSizeBytes) {
             Slog.w(TAG, "RemoteViews too large on pkg: " + pkg + " tag: " + tag + " id: " + id
